refactor: dedupe framework metadata helpers
Move Angular and Nest framework-marking casts into package-local helpers, clarify the server internal mutator comment, and inline the now-unneeded private client factory helpers in web and server.
